examples/requirements.txt
fastapi
uvicorn
python-dotenv
openai
langdag
pydantic